speculation-rules/prefetch/initiators-a-element.sub.https.html

<!DOCTYPE html>
<meta name="variant" content="?cross-site">
<meta name="variant" content="?same-site">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.sub.js"></script>
<script>
  // In https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigate,
  // `sourceDocument` (instead of `navigable`'s active document) should be
  // used as the referring document for prefetch.

  const { from_protocol, to_protocol } = Object.fromEntries(new URLSearchParams(location.search));
  promise_test(async t => {
    assert_implements(HTMLScriptElement.supports('speculationrules'), "Speculation Rules not supported");

    const win = await spawnWindow(t, { protocol: 'https' });

    const hostname =
      location.search === '?cross-site' ? '{{hosts[alt][www]}}' : undefined;
    const nextUrl = win.getExecutorURL({ protocol: 'https', hostname, page: 2 });

    await win.forceSinglePrefetch(nextUrl);

    await win.execute_script((url) => {
      window.executor.suspend(() => {
        const a = document.createElement('a');
        a.setAttribute('href', url);
        document.body.appendChild(a);
        a.click();
      });
    }, [nextUrl]);

    assert_equals(
        await win.execute_script(() => location.href),
        nextUrl.toString(),
        "expected navigation to reach destination URL");

    assert_prefetched(await win.getRequestHeaders());
  }, `<a>`);

  promise_test(async t => {
    assert_implements(HTMLScriptElement.supports('speculationrules'), "Speculation Rules not supported");

    const win = await spawnWindow(t, { protocol: 'https' });

    const hostname =
      location.search === '?cross-site' ? '{{hosts[alt][www]}}' : undefined;
    const nextUrl = win.getExecutorURL({ protocol: 'https', hostname, page: 2 });

    await win.forceSinglePrefetch(nextUrl);

    await win.execute_script((url) => {
      window.executor.suspend(() => {
        const a = document.createElement('a');
        a.setAttribute('href', url);
        a.setAttribute('target', '_blank');
        document.body.appendChild(a);
        a.click();
      });
    }, [nextUrl]);

    assert_equals(
        await win.execute_script(() => location.href),
        nextUrl.toString(),
        "expected navigation to reach destination URL");

    assert_prefetched(await win.getRequestHeaders());
  }, `<a target="blank">`);
</script>
